requirements.txt
atlasify==0.8.0
coverage==6.3.1
h5py==3.10.0
matplotlib==3.5.1
numpy==1.24.4
pandas[hdf5]==1.5.3
pre-commit==3.5.0
pydot==1.4.2
pytest-cov==3.0.0
pytest-randomly==3.11.0
pytest==7.0.1
pyyaml-include==1.3
scipy==1.10.1
tables==3.8.0
testfixtures==7.0.0
palettable==3.3.0
atlas-ftag-tools==0.1.19
puma-hep